fixed bug in make_master_calib() when scaling with texp is not possible for lamp_config='both'; made temp-bg files in process_science_images() single precision

+++ process_scripts.py
@@ -542,7 +542,7 @@ def process_science_images(imglist, P_id, chipmask, mask=None, stripe_indices=No
                 # fit background
                 bg_coeffs, bg_img = fit_background(bg, clip=10, return_full=True, timit=timit)
                 # save background image to temporary file for re-use later (when reducing the next file of this sublist)
-                pyfits.writeto(path + 'temp_bg_' + lamp_config + '.fits', bg_img, overwrite=True)
+                pyfits.writeto(path + 'temp_bg_' + lamp_config + '.fits', np.float32(bg_img), overwrite=True)
             else:
                 # no need to re-compute background, just load it from file
                 print('Loading background image for this epoch and lamp configuration...')
@@ -578,7 +578,7 @@ def process_science_images(imglist, P_id, chipmask, mask=None, stripe_indices=No
                 # fit background
                 bg_coeffs, bg_img = fit_background(bg, clip=10, return_full=True, timit=timit)
                 # save background image to temporary file for re-use later (when reducing the next file of this sublist)
-                pyfits.writeto(path + 'temp_bg_' + lamp_config + '.fits', bg_img, overwrite=True)
+                pyfits.writeto(path + 'temp_bg_' + lamp_config + '.fits', np.float32(bg_img), overwrite=True)
             else:
                 # no need to re-compute background, just load it from file
                 print('Loading background image for this epoch and lamp configuration...')
